# Optimizations for GPU memory