--- a/include/lbann/layers/activations/identity.hpp
+++ b/include/lbann/layers/activations/identity.hpp
@@ -31,38 +31,26 @@
 
 namespace lbann {
 
-/** Identity activation function. */
+/** Identity layer. */
 template <data_layout T_layout, El::Device Dev>
 class identity_layer : public activation_layer {
- public:
+public:
   identity_layer(lbann_comm *comm) : activation_layer(comm) {}
   identity_layer* copy() const override { return new identity_layer(*this); }
   std::string get_type() const override { return "identity"; }
   data_layout get_data_layout() const override { return T_layout; }
   El::Device get_device_allocation() const override { return Dev; }
 
-  void setup_gpu() override {
-    activation_layer::setup_gpu();
-#ifdef HYDROGEN_HAVE_CUB
-    // Set matrices to use CUB GPU memory pool
-    // Note: During each iteration, matrices are resized to the
-    // mini-batch size and cleared to obtain a matrix view. To avoid
-    // expensive GPU memory allocations and deallocations, we use
-    // CUB's GPU memory pool.
-    if (using_gpus()) {
-      get_local_activations().SetMemoryMode(1);
-      get_local_error_signals().SetMemoryMode(1);
-    }
-#endif
-  }
+protected:
 
-  void fp_compute() override {
+  void fp_setup_outputs(El::Int mini_batch_size) override {
     El::LockedView(get_activations(), get_prev_activations());
   }
-
-  void bp_compute() override {
+  void bp_setup_gradient_wrt_inputs(El::Int mini_batch_size) override {
     El::LockedView(get_error_signals(), get_prev_error_signals());
   }
+  void fp_compute() override {}
+  void bp_compute() override {}
 
 };