--- a/include/lbann/layers/layer.hpp
+++ b/include/lbann/layers/layer.hpp
@@ -105,16 +105,15 @@ class Layer {
   virtual std::string get_topo_description() const;
 
   /** Forward propagation step.
-   *  Apply a mathematical operation to the previous activations to
-   *  obtain the activations.
+   *  Apply a mathematical operation to input tensors to obtain output
+   *  tensors.
    */
   virtual void forward_prop();
   /** Backward propagation step.
-   *  Given the objective function gradients w.r.t. the activations
-   *  (the previous error signals), compute the gradients w.r.t. the
-   *  previous activations (the error signals) and w.r.t. the
-   *  weights. This is essentially an application of the chain
-   *  rule.
+   *  Given the objective function gradients w.r.t. the output
+   *  tensors, compute the gradients w.r.t. the input tensors and
+   *  w.r.t. the weights. This is essentially an application of the
+   *  chain rule.
    */
   virtual void back_prop();
   /** Update step.
@@ -207,6 +206,10 @@ class Layer {
   std::string get_child_names() const { return get_layer_names(m_child_layers); }
   std::string get_parent_names() const { return get_layer_names(m_parent_layers); }
 
+  // ===========================================================
+  // Layer pointer manipulation functions
+  // ===========================================================
+
   /** Add a parent layer.
    *  Does nothing if parent is a null pointer, the same layer, or
    *  already a parent.
@@ -232,6 +235,10 @@ class Layer {
   /** Set list of pointers to other layers. */
   virtual void set_layer_pointers(std::vector<Layer*> layers);
 
+  // ===========================================================
+  // Weights access functions
+  // ===========================================================
+
   /** Get references to weights. */
   inline std::vector<weights*>& get_weights() { return m_weights; }
   /** Get references to weights. (const) */
@@ -240,6 +247,10 @@ class Layer {
   inline void set_weights(std::vector<weights*> w) { get_weights() = w; }
   /** Replace weights with another Layer's weights*/
   void replace_weights(Layer* other_layer);
+
+  // ===========================================================
+  // Tensor dimension access functions
+  // ===========================================================
 
   /** Get dimensions of an input tensor.
    *  E.g. get the dimensions of a "previous activations tensor" or
@@ -261,6 +272,10 @@ class Layer {
    *  "neurons."
    */
   int get_output_size(int output_index = 0) const;
+
+  // ===========================================================
+  // Tensor access functions
+  // ===========================================================
 
   /** Get activation tensor. */
   AbsDistMat& get_activations(int child_index = 0);
@@ -290,11 +305,119 @@ class Layer {
   /** Get reference to LBANN communicator. */
   lbann_comm* get_comm() const { return m_comm; }
 
+  // ===========================================================
+  // Freeze management functions
+  // ===========================================================
+
   void freeze();
   void unfreeze();
   bool is_frozen() const;
 
  protected:
+
+  /** Set dimensions of an output tensor.
+   *  E.g. set the dimensions of an "activations tensor" or the
+   *  "neuron dimensions."
+   */
+  void set_output_dims(std::vector<int> dims, int output_index = 0);
+
+  // ===========================================================
+  // Setup helper functions
+  // ===========================================================
+
+  /** Setup layer pointers.
+   *  Called by the 'setup' function. Pointers to parent/child layers
+   *  are assumed to be already initialized.
+   */
+  virtual void setup_pointers();
+  /** Setup tensor dimensions
+   *  Called by the 'setup' function. If there are any input tensors,
+   *  the base method sets all uninitialized output tensor dimensions
+   *  equal to the first input tensor dimensions.
+   */
+  virtual void setup_dims();
+  /** Setup distributed matrices.
+   *  Called by the 'setup' function. Each column of these distributed
+   *  matrices is interpreted as the flattened tensor for a mini-batch
+   *  sample. The matrices themselves are constructed by calling the
+   *  'construct_matrix' function. If any matrices have already been
+   *  setup, they are destroyed and reinstantiated.
+   */
+  virtual void setup_matrices(const El::Grid& grid);
+  /** Construct distributed matrix.
+   *  Called by the 'setup_matrices' function. 'type' is one of the
+   *  following: "input", "output", "gradient_wrt_output",
+   *  "gradient_wrt_input".
+   */
+  virtual std::unique_ptr<AbsDistMat> construct_matrix(const El::Grid& grid,
+                                                       std::string type,
+                                                       El::Int index);
+  /** Setup layer data.
+   *  Called by the 'setup' function. Memory is allocated for
+   *  distributed matrices.
+   */
+  virtual void setup_data();
+  /** Setup GPU objects.
+   *  Called by the 'setup' function if the layer is on GPUs.
+   */
+  virtual void setup_gpu() {}
+
+  // ===========================================================
+  // Forward prop step helper functions
+  // ===========================================================
+
+  /** Setup input tensors.
+   *  Called by the 'forward_prop' function. Each input tensor is
+   *  setup as a view or copy of the corresponding parent layer's
+   *  output tensor.
+   */
+  virtual void fp_setup_inputs(El::Int mini_batch_size);
+  /** Setup output tensors.
+   *  Called by the 'forward_prop' function. Each output tensor is
+   *  resized to match the mini-batch size.
+   */
+  virtual void fp_setup_outputs(El::Int mini_batch_size);
+  /** Apply layer operation.
+   *  Called by the 'forward_prop' function. Given the input tensors,
+   *  the output tensors are populated with computed values.
+   */
+  virtual void fp_compute() = 0;
+
+  // ===========================================================
+  // Back prop step helper functions
+  // ===========================================================
+
+  /** Setup gradient w.r.t. output tensors.
+   *  Called by the 'back_prop' function. Each gradient w.r.t. output
+   *  tensor is setup as a view or copy of the corresponding child
+   *  layer's gradient w.r.t. input tensor.
+   */
+  virtual void bp_setup_gradient_wrt_outputs(El::Int mini_batch_size);
+  /** Setup gradient w.r.t. input tensors.
+   *  Called by the 'back_prop' function. Each gradient w.r.t. input
+   *  tensor is resized to match the mini-batch size.
+   */
+  virtual void bp_setup_gradient_wrt_inputs(El::Int mini_batch_size);
+  /** Compute objective funciton gradients.
+   *  Called by the 'back_prop' function. Given the input, output, and
+   *  gradient w.r.t. output tensors, the gradient w.r.t. input
+   *  tensors are populated with the computed values and the gradients
+   *  w.r.t. the weights are sent to the appropriate optimizers.
+   */
+  virtual void bp_compute();
+
+  // ===========================================================
+  // Update step helper functions
+  // ===========================================================
+
+  /** Perform the computation for the update step.
+   *  Returns false if the layer must reset for a new training epoch.
+   */
+  virtual bool update_compute() { return true; }
+
+  // ===========================================================
+  // Protected class members
+  // ===========================================================
 
   /** Reference to LBANN communicator. */
   lbann_comm *m_comm;
@@ -319,70 +442,6 @@ class Layer {
   /** Reference to model managing this layer. */
   model *m_model = nullptr;
 
-  /** Set dimensions of an output tensor.
-   *  E.g. set the dimensions of an "activations tensor" or the
-   *  "neuron dimensions."
-   */
-  void set_output_dims(std::vector<int> dims, int output_index = 0);
-
-  /** Setup data for forward propagation.
-   *  Base method gets previous activations from parent layers and
-   *  resizes activations for the current mini-batch size.
-   */
-  virtual void fp_setup_data(int mini_batch_size);
-  /** Setup data for forward propagation.
-   *  Base method gets previous error signals from child layers and
-   *  resizes error signals for the current mini-batch size.
-   */
-  virtual void bp_setup_data(int mini_batch_size);
-
-  /** Setup layer pointers.
-   *  Called by the 'setup' function. Pointers to parent/child layers
-   *  are assumed to be already initialized.
-   */
-  virtual void setup_pointers();
-  /** Setup tensor dimensions
-   *  Called by the 'setup' function. If there are any input tensors,
-   *  the base method sets all uninitialized output tensor dimensions
-   *  equal to the first input tensor dimensions.
-   */
-  virtual void setup_dims();
-  /** Setup distributed matrices.
-   *  Called by the 'setup' function. Each distributed matrix is
-   *  constructed by calling the 'construct_matrix' function. If any
-   *  matrices have already been setup, they are destroy and
-   *  reinstantiated.
-   */
-  virtual void setup_matrices(const El::Grid& grid);
-  /** Construct distributed matrix.
-   *  Called by the 'setup_matrices' function. 'type' is one of the
-   *  following: "input", "output", "gradient_wrt_output",
-   *  "gradient_wrt_input".
-   */
-  virtual std::unique_ptr<AbsDistMat> construct_matrix(const El::Grid& grid,
-                                                       std::string type,
-                                                       El::Int index);
-  /** Setup layer data.
-   *  Called by the 'setup' function. Resizes matrices to the
-   *  appropriate dimensions.
-   */
-  virtual void setup_data();
-  /** Setup GPU objects.
-   *  Called by the 'setup' function if the layer is on GPUs.
-   */
-  virtual void setup_gpu() {}
-
-  /** Perform the computation for the forward propagation step. */
-  virtual void fp_compute() = 0;
-  /** Perform the computation for the backward propagation step.
-   *  The base implementation sets all error signals to zero.
-   */
-  virtual void bp_compute();
-  /** Perform the computation for the update step.
-   *  Returns false if the layer must reset for a new training epoch.
-   */
-  virtual bool update_compute() { return true; }
-
   /** Avoid back prop if frozen */
   bool m_frozen;
 
@@ -405,6 +464,19 @@ class Layer {
 
  private:
 
+  // ===========================================================
+  // Private access functions
+  // ===========================================================
+
+  /** Get activation tensor corresponding to child layer. */
+  const AbsDistMat& get_activations(const Layer& child) const;
+  /** Get error signal tensor corresponding to parent layer. */
+  const AbsDistMat& get_error_signals(const Layer& parent) const;
+
+  // ===========================================================
+  // Private class members
+  // ===========================================================
+
   /** Dimensions of output tensors. */
   std::vector<std::vector<int>> m_output_dims_list;
 
@@ -424,11 +496,6 @@ class Layer {
    *  Each matrix column corresponds to a flattened mini-batch sample.
    */
   std::vector<std::unique_ptr<AbsDistMat>> m_gradient_wrt_inputs;
-
-  /** Get activation tensor corresponding to child layer. */
-  const AbsDistMat& get_activations(const Layer& child) const;
-  /** Get error signal tensor corresponding to parent layer. */
-  const AbsDistMat& get_error_signals(const Layer& parent) const;
 
 };