… include/lbann/layers/transform/concatenation.hpp
… include/lbann/layers/transform/slice.hpp
… include/lbann/layers/transform/stop_gradient.hpp
… src/layers/layer.cpp
… src/proto/factories/layer_factory.cpp