### Summary
Miscellaneous work (mostly in the base layer class) to optimize our GPU memory footprint without losing performance. See #345.

### Changelog
- GPU layers allocate tensor memory with the CUDA API, not the CUB memory pool.
- Layers that output tensor views (e.g. identity, reshape, split, sum) do not allocate memory for them.
- Added a virtual `construct_matrix()` function in the base layer class that derived classes can override to change matrix distributions and memory modes.
- Split up `fp_setup_data()` in the base layer class into `fp_setup_inputs()` and `fp_setup_outputs()`. Derived classes can override `fp_setup_outputs()` to avoid having to allocate unnecessary memory.
- Split up `bp_setup_data()` in the base layer class into `bp_setup_gradient_wrt_outputs()` and `bp_setup_gradient_wrt_inputs()`.
- `setup_data()` in the base layer class calls `fp_setup_inputs()` and friends to reduce code duplication.
- Fixed some bugs involving matrix alignment.
- Dropout no longer outputs views during evaluation. This gave a tiny performance improvement during evaluation at the expense of code complexity. Something something premature optimization.
- Slice and concatenation layers output views during forward prop and back prop, when possible.

### Results
Experiment results with Resnet-50 (1 Surface gpgpu node, 4 procs/node; mini-batch size 32; ImageNet-10):

Commit | GPU 0 usage at epoch 0 | GPU 0 usage at epoch 9 | Epoch 0 time | Average epoch time (excluding epoch 0)
--------|--------------|--------------|----------|--------------
3982aaa (develop) | 4.68 GiB | 5.79 GiB | 89.8s | 70.9s
1c5f057 | 3.93 GiB | 4.71 GiB | 84.7s | 70.6s
fa6bdf4 | 3.16 GiB | 4.26 GiB | 97.3s | 71.1s

I suspect the memory difference between epoch 0 and epoch 9 is because the convolution workspace is allocated during the first convolution layer's forward prop.

### Validation
- Gradient checks on model_zoo/tests/model_mnist_conv_graph.prototext come back clean.
- Reasonable results with AlexNet on ImageNet-10 (15.2%/59.5% top-1/top-5 training accuracy, 30%/76.2% top-1/top-5 validation accuracy, 12.5s training time for epoch 0).