@ndryden 2018-07-20
What does the memory usage look like at epoch 1? I mostly want to make sure that the increase in memory usage really is just allocating convolution workspaces and other similar things, and that we don't have a memory leak.

Can we run some larger-scale performance tests (multi-node, ImageNet-1k) to make sure that there are no performance regressions?