@timmoon10 2018-07-20
The memory usage at epoch 1 is the same as at epoch 9.