@timmoon10 2018-07-20
Experiment results with Resnet-50 (16 Pascal nodes, 2 procs/node; mini-batch size 256; ImageNet-1K):

Commit | GPU 0 usage at epoch 0 | GPU 0 usage at epoch 1 | Epoch 0 time | Epoch 1 time
--------|--------------|--------------|----------|--------------
325e5b7c (develop) | 4.65 GiB | 5.78 GiB | 21.7 min | 19.7 min
92896429 | 3.25 GiB | 4.37 GiB | 24.2 min | 22 min