@ndryden 2018-07-20
I'm a little concerned that this appears to result in a ~10% performance hit. Do we know where that is coming from? (It shouldn't be dropout, since that's not used in ResNet-50.)