@timmoon10 2018-07-20
Repeating the last experiment a few times, it looks like the performance hit is real:

 Branch | Epoch 0 time | Epoch 1 time
-------|----------|----------
develop | 1112.15s | 1131.47s
develop | 1147.27s | 1061.5s
develop | 1146.46s | 1061.5s
gpu-memory-optimization | 1652.79s | 1491.19s
gpu-memory-optimization | 1571.84s | 1446.91s

I'll need to do some profiling to figure out what's going on.