@timmoon10 2018-07-24
I've run the experiment several more times, but made sure to alternate between the develop branch and this PR:

Branch | Epoch 0 time | Epoch 1 time
-------|----------|----------
gpu-memory-optimization | 1430.66s | 1001.56s
develop | 1430.24s | 1001.55s
gpu-memory-optimization | 1523.79s | 1121.02s
develop | 1151.16s | 1030.9s
gpu-memory-optimization | 1072.83s | 1067.02s
develop | 1065.87s | 1035.5s
gpu-memory-optimization | 1045.94s | 1153.23s
develop | 1138.97s | 1151.65s

On average, this PR has an epoch 1 time that's ~30s slower. However, the sample standard deviation in these runs is ~60s. I didn't see anything unexpected when profiling with nvprof, so I'm inclined to blame the performance regression on weird variability on Pascal's network.