@ndryden 2018-07-24
This agrees with my experiments.